refactor: extract downgrade info to a component
